+++ src/main/java/com/nexmo/client/verify/endpoints/VerifyEndpoint.java
@@ -24,6 +24,7 @@ package com.nexmo.client.verify.endpoints;
 import com.nexmo.client.HttpWrapper;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.NexmoResponseParseException;
+import com.nexmo.client.auth.SignatureAuthMethod;
 import com.nexmo.client.auth.TokenAuthMethod;
 import com.nexmo.client.legacyutils.XmlParser;
 import com.nexmo.client.verify.VerifyRequest;
@@ -41,7 +42,7 @@ import java.io.UnsupportedEncodingException;
 import java.util.Locale;
 
 public class VerifyEndpoint extends AbstractMethod<VerifyRequest, VerifyResult> {
-    private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{TokenAuthMethod.class};
+    private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{SignatureAuthMethod.class, TokenAuthMethod.class};
 
     private static final String DEFAULT_URI = "https://api.nexmo.com/verify/xml";