+++ src/test/java/com/nexmo/client/numbers/UpdateNumberEndpointTest.java
@@ -30,9 +30,7 @@ import org.junit.Test;
 
 import java.util.Map;
 
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
+import static org.junit.Assert.*;
 
 public class UpdateNumberEndpointTest {
     private UpdateNumberEndpoint endpoint;